requirement.txt
asgiref==3.2.7
audioread==2.1.8
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
Django==3.0.5
django-extensions==2.2.9
django-sslserver==0.22
hmm==0.0.1
hmmlearn==0.2.3
joblib==0.15.1
librosa==0.7.2
llvmlite==0.31.0
numba==0.48.0
numpy==1.18.1
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.32
Pillow==7.1.1
pycparser==2.20
python-speech-features==0.6
pytz==2019.3
resampy==0.2.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.14.0
SoundFile==0.10.3.post1
sqlparse==0.3.1
threadpoolctl==2.1.0
Werkzeug==1.0.1
